Updated pins_arduino.h for ESP32 POE, POE-ISO and Gateway
Due to the latest changes in the ethernet library/examples (https://github.com/espressif/arduino-esp32/pull/9242) full package of definitions of the ethernet macros is for each board that has non-default values.
For POE and POE-ISO is added a preprocessor condition #if defined BOARD_HAS_PSRAM due to our specific hardware because GPIO 16 and 17 are in use when PSRAM is enabled.