Update CameraWebServer.ino
Updated the loop function of CameraWebServer example. Currently this software will not try to re-connect to the network if it lost once. There may be a lot of reasons - wifi router lost electricity, some noise in the radio space and much more. Now, the loop function will check for the connection, it is lost - will try to re-connect. Once re-conencted, will run the web server again.