fix(esp32-hal-ledc): Clear bit in used_channels bitmap
When we attach pin to LEDC channel we set bit in used_channels bitmap variable.
But used_channels bit was not cleared on detach,
as result simply repeating attach/detach around 16 times
will falsely use all available channels and
emit error that channel is not available.

This happen for example in IRSenderESP32.cpp.
This patch fixes this bug.

Signed-off-by: Denys Fedoryshchenko <denys.f@collabora.com>