fix(boards): Correct PID port property in "Bee Data Logger" board def
VID and PID port property values are hexadecimal literals and thus use the standard "0x" prefix.

Previously, this prefix was missing from the PID port property association in the definition of the
board definition of "Bee Data Logger" board. This typo caused ports having the 0x303a, 0x815C
VID/PID pair to not be identified as a "Bee Data Logger" board as intended.