fix(boards): Correct property names for PID port property associations
These properties are intended to associate values of the PID port property with "Bee" board
definitions. A typo caused them to instead associatiate the values with the VID port property. The
result was that ports with this VID/PID pair were not identified by the Arduino development
software as the board model as intended.