fix(boards): Correct typos in port prop. associations for "Bee" boards (#10276)
* fix(boards): Correct PID port property in "Bee Data Logger" board def

VID and PID port property values are hexadecimal literals and thus use the standard "0x" prefix.

Previously, this prefix was missing from the PID port property association in the definition of the
board definition of "Bee Data Logger" board. This typo caused ports having the 0x303a, 0x815C
VID/PID pair to not be identified as a "Bee Data Logger" board as intended.

* fix(boards): Correct property names for PID port property associations

These properties are intended to associate values of the PID port property with "Bee" board
definitions. A typo caused them to instead associatiate the values with the VID port property. The
result was that ports with this VID/PID pair were not identified by the Arduino development
software as the board model as intended.