fix(boards): Remove port property assoc. for non-unique VID/PID pairs (#10271)
The Arduino boards platform framework allows properties of a port to be associated with a board
definition. The Arduino development software will identify a port having the associated properties
as that board.

This should only be done for properties that are unique to the ports produced by that board model.
In cases where a board model does not produce a port with properties unique to that model, it is
irresponsible to associate those properties with the board definition as this will cause other
models to be inappropriately identified.

The authors of these board definitions associated them with non-unique USB VID/PID pairs so those
associations must be removed. The 303a:1001 VID/PID pair is of the hardware CDC serial port of the
native USB ESP32 chips. The others are stock VID/PID pairs of general purpose USB to serial bridge
chips that are used on many different board models.