Support HTTP 204 (#11408)
HTTP 204 is a successful return code which indicates No Content. While it's appropriate to return a 304 if the server has content for a device but it hasn't change, it is more accurate for a server to return a 204 if it simply doesn't have any firmware files for a particular device.

Co-authored-by: Me No Dev <me-no-dev@users.noreply.github.com>