feat(usb): allow the MIDI constructor to define a device name
This PR will allow the user to modify the MIDI USB Device Name (device descriptor). Fomer API continues to be valid and the default device name is "TinyUSB MIDI". 

The user can change the device name by passing a string to the constructor, declaring it like this `USBMIDI myMIDI("MyDeviceName");