refactor(ble): Pass pointer to const data to setValue()
There is no technical need to pass non-const data there.

Also it is a common practice to pass const data to the functions
with write semantics, as they not supposed to modify the data.

This allows to pass the data from read-only sources without
removing the constness what is not always safe.